Shockingly successful
only one error i've noticed.
Loading song versions is iffy.
Because i save files based only off the name of the folder they are in, a song version (for example tutorial_hard)
will save as tutorial.json, then trying to load the default tutorial will load the json which contains the _hard notes.
Make it save files with respect to the filepath, not the folder.