Questionable hack
Added constructor override to note event info that lets me store the info of a Tempo change event in the same class. saves filespace, and time implementing  a seperate class but is sure to cause an issue eventually. Just be sure to check the start and note number of all NEI if you are using their values.