Small patches
no more input from midi after the song is done, fixed scores saving to the Kiboard/songs  folder instead of the song specific folder

REPLAYS KINDA BROKEN, it tracks the first note you hit to begin the song playing as the first note causing a really big offset for the rest of them from the music. Remove the first note.