implemented UI Image Tags and Custom Song Image Tags
Add the RuntimeImageLoader script to any object with a rawImage component.
set isSongSpecific to true if it is based on the selected song, or false if it is a general image in the Kiboard/Images/ directory.
Type the Tag or prefix for the image into imagePrefix. it will assign on start.

NEXT, FIX THE TEMPO CHANGING DESYNC ISSUE, AND CONTINUE WORK ON THE SONG EDITOR.
Side note, can almost certainly use the replay function to add a recording part to the songeditor.