fixed weird error with tending to round up with notes, and began on editor UI
Thinking of simple ways to do this, i kinda want to be able to select 2 -3 tools for l/r/m mouse button that store an Enum.
then can use that enum to determine each click's behaviour.

the ideal way of doing this would be to make an action/event for the mouse clicks and changes and subscribe/unsubscribe methods to those events at the correct times, but idk how to do that. will look into it.