Added togglegroup for selecting vSnap values
common ones have their own button (1,2,3,4,8,16), or you can type your own with the "#" button.

Work on note inspector, camera scrolling, and playtesting.

For playtesting there are two ways of potentially doing it:
ideally we load gamescene88 for it so it is a perfect 1-1 of how the notes will play and appear.
could maybe async load the scene and keep the editor loaded? maybe put the scene under a parent object then dontdestroy on load?. Find a way to pass the entire scene through gamescene88 then get it back.

Option 2, rewrite a songeditor only version to work with the completely vertical piano. would be nice for just testing and i think i should do this anyways and give the player the option between the two.